+++ packages/core/src/git/index.ts
@@ -2,3 +2,4 @@ export * from './url.js';
 export * from './files.js';
 export * from './fetcher.js';
 export * from './tar.js';
+export * from './tarball.js';